Expose the app-wide defaults, and make 0 redirects mean unlimited
Three things, all about defaults that were either hidden or backwards.

Maximum redirects read "0 uses 10", which is the wrong way round: the
field showed a number that wasn't the answer, and there was no way to
say "follow without limit". It now shows the real cap and 0 means no cap
at all -- only the request's timeout ends a redirect loop.

domain.Options.MaxRedirects becomes a *int so that "never set" and
"explicitly 0" stay distinguishable. Without it, every request saved
before today -- where the key is simply absent -- would have started
following redirects forever. Codegen resolves the pointer once and each
language spells "no cap" its own way: curl --max-redirs -1, PowerShell's
int ceiling, and requests' own limit raised, since neither of the latter
two has an unlimited. The default cap is now always emitted rather than
left implicit, the same rule the timeout already followed.

internal/settings holds the four values that were constants in
internal/httpengine -- the request timeout, the redirect cap, and the
sizes above which a response stops being shown inline and stops being
read at all. They persist as settings.yaml in the workspace, so a
workspace shared through git carries them, and they are edited in
Settings > Workspace rather than by hand. Go clamps whatever arrives and
returns what it kept, so a ceiling below the inline threshold is raised
instead of truncating a body the pane was about to render. The first two
seed every new request's Options, so changing the workspace default
changes what new requests get.

The InfoTip popup is now position: fixed. Every place it is used sits
inside a scroll container -- .request-pane, .settings-body -- and
overflow-y: auto computes overflow-x with it, so an absolutely
positioned popup was clipped on all four sides. It measures the button
and places itself in viewport coordinates, flipping to stay on screen,
which also retires the align prop the caller had to guess at.

Not verified by clicking: synthetic mouse input does not reach the
WebView2 surface, so the popup could not be opened from a script. The
CSS guarantee is that a fixed element is not clipped by an ancestor's
overflow; worth a look by hand.

+++ internal/wailsapp/app.go
@@ -104,7 +104,7 @@ func (a *App) ExecuteRequest(collectionID, itemID, environmentID string) (*httpe
 // the shared shape ExecuteRequest and GetCachedResponse both hand the
 // frontend for an oversized response.
 func (a *App) trimLargeBody(itemID string, resp *httpengine.Response) {
-	if len(resp.Body) <= httpengine.LargeResponseThreshold {
+	if int64(len(resp.Body)) <= httpengine.LargeResponseThreshold {
 		return
 	}
 	if path, err := a.responseCacheBodyPath(itemID); err == nil {
